client/GitHubRequest: abstracted github request component
- for use in ToolExplanation and UseAccessToken
- refactored usage in UseAccessToken
